Push notifications-like post

+++ client/src/hooks/like/useDeleteLike.ts
@@ -6,25 +6,37 @@ import { IPost } from "@/types/IPost";
 import { useParams } from "@tanstack/react-router";
 import { IGetAllPostsResponse } from "@/data/post/getAllPosts";
 import { IGetPostsByUsernameResponse } from "@/data/post/getPostsByUsername";
-import { likeSchema } from "@/validation/likeSchema";
+import { deleteLikeSchema } from "@/validation/likeSchema";
 
-const dislikePost = async (postId: string) => {
-  const parsed = likeSchema.safeParse({ postId });
+interface IDeleteLikeByPostIdResponse {
+  postId: string;
+}
+
+export interface IDeleteLikePayload {
+  postId: string;
+  author_id: string;
+}
+
+const dislikePost = async ({ postId, author_id }: IDeleteLikePayload) => {
+  const parsed = deleteLikeSchema.safeParse({ params: { postId }, body: { author_id } });
   if (!parsed.success) {
     throw new Error(`${parsed.error.issues[0].message} - ${parsed.error.issues[0].path}`);
   }
-  const url = `/api/likes/${parsed.data.postId}`;
-  const res = await apiClient.delete(url);
+  const url = `/api/likes/${parsed.data.params.postId}`;
+  const res = await apiClient.delete(url, { data: { author_id: parsed.data.body.author_id } });
   return res.data;
 };
 
-export const useDeleteLike = ({ postId }: { postId: string }) => {
+export const useDeleteLike = () => {
   const { username } = useParams({ strict: false });
   const queryClient = useQueryClient();
   return useMutation({
-    mutationFn: () => dislikePost(postId),
-    onError: (error) => toast.error(error.message || "Failed to dislike post."),
-    onSuccess: () => {
+    mutationFn: dislikePost,
+    onError: (error) => {
+      console.log(`${error.message} - delete like`);
+      toast.error("Failed to dislike post or post doesn't exist.");
+    },
+    onSuccess: (data: IDeleteLikeByPostIdResponse) => {
       queryClient.setQueryData(postKeys.list(), (oldData: InfiniteData<IGetAllPostsResponse>) => {
         // Safeguard against undefined oldData
         if (!oldData || !oldData.pages) return oldData;
@@ -36,7 +48,7 @@ export const useDeleteLike = ({ postId }: { postId: string }) => {
           return {
             ...page,
             data: page.data.map((post) => {
-              if (post && post.id === postId) {
+              if (post && post.id === data.postId) {
                 const currentLikeCount = Number(post.like_count) || 0;
                 return {
                   ...post,
@@ -56,7 +68,7 @@ export const useDeleteLike = ({ postId }: { postId: string }) => {
         };
       });
 
-      queryClient.setQueryData(postKeys.detail(postId), (oldData: IPost) => {
+      queryClient.setQueryData(postKeys.detail(data.postId), (oldData: IPost) => {
         if (!oldData) {
           console.warn("Old data for post detail is undefined. Skipping update.");
           return oldData; // Safeguard for undefined oldData
@@ -76,7 +88,7 @@ export const useDeleteLike = ({ postId }: { postId: string }) => {
             return {
               ...page,
               data: page.data.map((post) => {
-                if (post && post.id === postId) {
+                if (post && post.id === data.postId) {
                   const currentLikeCount = Number(post.like_count) || 0;
                   return {
                     ...post,
